feat: move dialogue into features and exclude preformated text as new component
